Dogfood client HMR helpers with browser tests
